refactor: 노트 데이터에서 본문(content)내용만 받아오도록 Note 엔티티에서 title 필드 제거
