+++ src/entities/note/model/types.ts
@@ -1,6 +1,5 @@
 export interface Note {
   id: number;
-  title: string;
   content: string;
   studyGoalId: number;
   todoId: number;
