+++ src/features/edit-note/model/types.ts
@@ -1,5 +1,4 @@
 export interface UpdateNoteRequest {
   id: number;
-  title: string;
   content: string;
 }
