+++ src/features/edit-note/ui/NoteEditor.tsx
@@ -5,7 +5,7 @@ import { Note } from '@/entities/note/model/types';
 
 interface NoteEditorProps {
   initialNote?: Note;
-  onSubmit: (title: string, content: string) => void;
+  onSubmit: (content: string) => void;
   submitButtonText: string;
   isLoading?: boolean;
 }
@@ -16,43 +16,23 @@ export function NoteEditor({
   submitButtonText,
   isLoading = false,
 }: NoteEditorProps) {
-  const [title, setTitle] = useState('');
   const [content, setContent] = useState('');
 
   useEffect(() => {
     if (initialNote) {
-      setTitle(initialNote.title);
       setContent(initialNote.content);
     }
   }, [initialNote]);
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
-    if (title.trim() && content.trim()) {
-      onSubmit(title.trim(), content.trim());
+    if (content.trim()) {
+      onSubmit(content.trim());
     }
   };
 
   return (
     <form onSubmit={handleSubmit} className="space-y-6">
-      <div>
-        <label
-          htmlFor="title"
-          className="mb-2 block text-sm font-medium text-gray-700"
-        >
-          제목
-        </label>
-        <input
-          type="text"
-          id="title"
-          value={title}
-          onChange={(e) => setTitle(e.target.value)}
-          className="w-full rounded-md border border-gray-300 px-3 py-2 focus:border-transparent focus:ring-2 focus:ring-blue-500 focus:outline-none"
-          placeholder="노트 제목을 입력하세요"
-          required
-        />
-      </div>
-
       <div>
         <label
           htmlFor="content"
@@ -74,7 +54,7 @@ export function NoteEditor({
       <div className="flex justify-end">
         <button
           type="submit"
-          disabled={isLoading || !title.trim() || !content.trim()}
+          disabled={isLoading || !content.trim()}
           className="rounded-md bg-blue-500 px-6 py-2 text-white transition-colors hover:bg-blue-600 disabled:cursor-not-allowed disabled:bg-gray-400"
         >
           {isLoading ? '처리 중...' : submitButtonText}
